# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spotigen API",
    "version": "1.0.0",
    "description": "API for Spotigen – ChatGPT plugin to fetch favourite tracks and manage Spotify playlists."
  },
  "servers": [
    {
      "url": "https://spotigen-chat-gpt-plugin-production.up.railway.app"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root",
        "responses": {
          "200": {
            "description": "Service is up",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },

    "/logo.png": {
      "get": {
        "summary": "Plugin Logo",
        "operationId": "getLogo",
        "responses": {
          "200": {
            "description": "PNG logo",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },

    "/.well-known/ai-plugin.json": {
      "get": {
        "summary": "Plugin Manifest",
        "description": "Return the plugin manifest used by ChatGPT.",
        "operationId": "getManifest",
        "responses": {
          "200": {
            "description": "Manifest JSON",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },

    "/openapi.json": {
      "get": {
        "summary": "OpenAPI Document (JSON)",
        "operationId": "getOpenapiJson",
        "responses": {
          "200": {
            "description": "OpenAPI spec in JSON",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },

    "/auth/login": {
      "get": {
        "summary": "OAuth Login",
        "operationId": "login",
        "responses": {
          "307": {
            "description": "Redirect to Spotify OAuth"
          }
        }
      }
    },

    "/auth/callback": {
      "get": {
        "summary": "OAuth Callback",
        "operationId": "callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Auth success" }
        }
      }
    },

    "/auth/refresh": {
      "get": {
        "summary": "Refresh Access Token",
        "operationId": "refresh",
        "responses": {
          "200": { "description": "New access token" }
        }
      }
    },

    "/top_tracks": {
      "get": {
        "summary": "Get Top Tracks",
        "operationId": "getTopTracks",
        "responses": {
          "200": {
            "description": "Five most‑played tracks",
            "content": {
              "application/json": { "schema": { "type": "array", "items": {} } }
            }
          }
        },
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/playlist": {
      "get": {
        "summary": "Get Playlist",
        "operationId": "getPlaylist",
        "parameters": [
          { "name": "name", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Playlist found", "content": { "application/json": { "schema": {} } } },
          "404": { "description": "Playlist not found" }
        },
        "security": [ { "HTTPBearer": [] } ]
      },
      "post": {
        "summary": "Create Playlist",
        "operationId": "createPlaylist",
        "parameters": [
          { "name": "name", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "public", "in": "query", "required": true, "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": { "description": "Playlist created", "content": { "application/json": { "schema": {} } } }
        },
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/playlist/{playlist_id}/tracks": {
      "get": {
        "summary": "Get Playlist Tracks",
        "operationId": "getPlaylistTracks",
        "parameters": [
          { "name": "playlist_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {"200": { "description": "Tracks list", "content": { "application/json": { "schema": {} } } }},
        "security": [ { "HTTPBearer": [] } ]
      },
      "post": {
        "summary": "Add Tracks To Playlist",
        "operationId": "addTracksToPlaylist",
        "parameters": [
          { "name": "playlist_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {"required": true, "content": {"application/json": {"schema": { "$ref": "#/components/schemas/TrackTitles" }}}},
        "responses": {"200": { "description": "Tracks added" }},
        "security": [ { "HTTPBearer": [] } ]
      },
      "delete": {
        "summary": "Remove Tracks From Playlist",
        "operationId": "removeTracksFromPlaylist",
        "parameters": [
          { "name": "playlist_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {"required": true, "content": {"application/json": {"schema": { "$ref": "#/components/schemas/TrackURIs" }}}},
        "responses": {"200": { "description": "Tracks removed" }},
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/search": {
      "get": {
        "summary": "Search tracks",
        "operationId": "searchTracks",
        "parameters": [
          {"name": "q", "in": "query", "required": true, "schema": {"type": "string"}},
          {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer"}}
        ],
        "responses": {"200": {"description": "Results"}},
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/recommendations": {
      "get": {
        "summary": "Recommendations",
        "operationId": "recommendations",
        "parameters": [
          {"name": "seed_tracks", "in": "query", "required": true, "schema": {"type": "array", "items": {"type": "string"}}},
          {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer"}}
        ],
        "responses": {"200": {"description": "Recommended tracks"}},
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/audio_features": {
      "get": {
        "summary": "Audio Features",
        "operationId": "audioFeatures",
        "parameters": [
          {"name": "track_ids", "in": "query", "required": true, "schema": {"type": "array", "items": {"type": "string"}}}
        ],
        "responses": {"200": {"description": "Features"}},
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/recent": {
      "get": {
        "summary": "Recently Played",
        "operationId": "recentTracks",
        "parameters": [
          {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer"}}
        ],
        "responses": {"200": {"description": "Recent"}},
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/top_tracks": {
      "get": {
        "summary": "User Top Tracks",
        "operationId": "topTracks",
        "parameters": [
          {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer"}},
          {"name": "time_range", "in": "query", "required": false, "schema": {"type": "string"}}
        ],
        "responses": {"200": {"description": "Top tracks"}},
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/top_artists": {
      "get": {
        "summary": "User Top Artists",
        "operationId": "topArtists",
        "parameters": [
          {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer"}},
          {"name": "time_range", "in": "query", "required": false, "schema": {"type": "string"}}
        ],
        "responses": {"200": {"description": "Top artists"}},
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/stats": {
      "get": {
        "summary": "User Stats",
        "operationId": "stats",
        "responses": {"200": {"description": "Stats"}},
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/queue": {
      "post": {
        "summary": "Queue Track",
        "operationId": "queue",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"track_uri": {"type": "string"}}, "required": ["track_uri"]}}}},
        "responses": {"200": {"description": "Queued"}},
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/devices": {
      "get": {
        "summary": "Playback Devices",
        "operationId": "devices",
        "responses": {"200": {"description": "Devices"}},
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/play": {
      "post": {
        "summary": "Start Playback",
        "operationId": "play",
        "requestBody": {"required": false, "content": {"application/json": {"schema": {"type": "object", "properties": {"track_uri": {"type": "string"}, "device_id": {"type": "string"}}}}}},
        "responses": {"200": {"description": "Playing"}},
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/pause": {
      "post": {
        "summary": "Pause Playback",
        "operationId": "pause",
        "requestBody": {"required": false, "content": {"application/json": {"schema": {"type": "object", "properties": {"device_id": {"type": "string"}}}}}},
        "responses": {"200": {"description": "Paused"}},
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/skip_next": {
      "post": {
        "summary": "Skip Next",
        "operationId": "skipNext",
        "requestBody": {"required": false, "content": {"application/json": {"schema": {"type": "object", "properties": {"device_id": {"type": "string"}}}}}},
        "responses": {"200": {"description": "Skipped"}},
        "security": [ { "HTTPBearer": [] } ]
      }
    },

    "/agent": {
      "post": {
        "summary": "Natural Language Proxy",
        "operationId": "agent",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"prompt": {"type": "string"}}, "required": ["prompt"]}}}},
        "responses": {"200": {"description": "Free form response"}},
        "security": [ { "HTTPBearer": [] } ]
      }
    }
  },

  "components": {
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "TrackTitles": {
        "type": "object",
        "required": ["titles"],
        "properties": {
          "titles": { "type": "array", "items": { "type": "string" } }
        }
      },
      "TrackURIs": {
        "type": "object",
        "required": ["track_uris"],
        "properties": {
          "track_uris": { "type": "array", "items": { "type": "string" } }
        }
      }
    }
  }
}